change default phase of Maven TIA IT mojo to PACKAGE
was: pre-instrumentation-tests, but the Spring Boot plugin also starts
the application under test in the same phase. Ordering of goals within
phases is unfortunately dependant on the order of the plugins in the POM
which we can't easily enforce to be correct. So we move our goal earlier
in the lifecycle to avoid potential incorrect orderings